--- apps/meteor/client/views/marketplace/AccordionLoading.tsx
@@ -1,15 +1,0 @@
-import { Box, Skeleton, Margins } from '@rocket.chat/fuselage';
-import type { FC } from 'react';
-import React from 'react';
-
-const AccordionLoading: FC = () => (
-	<Box w='full' alignSelf='center'>
-		<Margins block={2}>
-			<Skeleton variant='rect' width='100%' height='x80' />
-			<Skeleton variant='rect' width='100%' height='x80' />
-			<Skeleton variant='rect' width='100%' height='x80' />
-		</Margins>
-	</Box>
-);
-
-export default AccordionLoading;
